+++ packages/ra-directus/src/index.ts
@@ -1,7 +1,13 @@
-export * from './addRefreshAuthToAuthProvider';
-export * from './addRefreshAuthToDataProvider';
+import {
+    addRefreshAuthToAuthProvider as addRefreshAuthToAuthProviderCore,
+    addRefreshAuthToDataProvider as addRefreshAuthToDataProviderCore,
+} from 'ra-core';
+
 export * from './directusAuthProvider';
 export * from './directusDataProvider';
 export * from './directusHttpClient';
 export * from './directusRefreshAuthToken';
 export * from './getDirectusProviders';
+
+export const addRefreshAuthToAuthProvider = addRefreshAuthToAuthProviderCore;
+export const addRefreshAuthToDataProvider = addRefreshAuthToDataProviderCore;