+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0
+
+-   Remove `addRefreshAuthToDataProvider` and `addRefreshAuthToAuthProvider` methods that are included in `ra-core` since version `4.9.0`. 
+
+To avoid breaking changes, we re-export the functions from `ra-core`.
+
 ## 1.0.1
 
 -   Fix `dataProvider` and `authProvider` ignore the provided storage